Refactored folder structure/naming to be consistent with cdk-patterns, added image for README.md
